+++ docs/src/components/MobileNavContext.ts
@@ -4,10 +4,6 @@ import type { Drawer } from '@base-ui/react/drawer';
 
 export const MobileNavContext = React.createContext<{
   handle: Drawer.Handle<unknown> | null;
-  searchValue: string;
-  setSearchValue: (value: string) => void;
 }>({
   handle: null,
-  searchValue: '',
-  setSearchValue: () => undefined,
 });
